+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 ## Next
 
 - test: testing release automation after sonatype migration
+- feat: add proxy to `PostHogConfig` ([#256](https://github.com/PostHog/posthog-android/issues/256))
+  
+```kotlin
+val config = PostHogAndroidConfig("...")
+config.proxy = Proxy(Proxy.Type.HTTP, InetSocketAddress("proxy.example.com",8080))
+```
 
 ## 3.17.0 - 2025-06-05